Refactor imports to use scanning.lib.js for ScanningProfile and CANDIDATES_PER_PAGE
